# LocadoraDeFilmes/Controllers/FilmesController.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using VideoLocadora.Domain.Features.Filmes;
using VideoLocadoraRepository.Features.Filmes;

namespace LocadoraDeFilmes.Controllers
{
    //Controller responsável para iteração com a entidade de FILME
    [Route("api/[controller]")]
    [ApiController]
    public class FilmesController : ControllerBase
    {
        private readonly AppDbContext _context;

        public FilmesController(AppDbContext context)
        {
            _context = context;
        }

        // GET: api/Filmes
        [HttpGet]
        public async Task<ActionResult<IEnumerable<Filme>>> GetFilmes()
        {
            return await _context.Filmes.ToListAsync();
        }

        // GET: api/Filmes/5
        [HttpGet("{id}")]
        public async Task<ActionResult<Filme>> GetFilme(int id)
        {
            var filme = await _context.Filmes.FindAsync(id);

            if (filme == null)
            {
                return NotFound();
            }

            return filme;
        }

        // PUT: api/Filmes/5
        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
        [HttpPut("{id}")]
        public async Task<IActionResult> PutFilme(int id, Filme filme)
        {
            if (id != filme.Id)
            {
                return BadRequest();
            }

            _context.Entry(filme).State = EntityState.Modified;

            try
            {
                await _context.SaveChangesAsync();
            }
            catch (DbUpdateConcurrencyException)
            {
                if (!FilmeExists(id))
                {
                    return NotFound();
                }
                else
                {
                    throw;
                }
            }

            return NoContent();
        }

        // POST: api/Filmes
        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
        [HttpPost]
        public async Task<ActionResult<Filme>> PostFilme(Filme filme)
        {
            _context.Filmes.Add(filme);
            await _context.SaveChangesAsync();

            return CreatedAtAction("GetFilme", new { id = filme.Id }, filme);
        }

        // MULTIPLE DELETE: api/Filmes/Ids (lista de string separado por virgula)
        [HttpDelete()]
        public async Task<IActionResult> DeleteMultiplosFilmes(string ids)
        {
            if (ids.Contains(","))
            {
                List<string> listaIdsParaDeletar = ids.Split(",").ToList();

                foreach (var id in listaIdsParaDeletar)
                {
                    var filme = await _context.Filmes.FindAsync(Convert.ToInt32(id));
                    if (filme == null)
                    {
                        return NotFound();
                    }

                    _context.Filmes.Remove(filme);
                    await _context.SaveChangesAsync();
                }
            }
            else
            {
                var filme = await _context.Filmes.FindAsync(Convert.ToInt32(ids));
                if (filme == null)
                {
                    return NotFound();
                }

                _context.Filmes.Remove(filme);
                await _context.SaveChangesAsync();
            }

            

            return NoContent();
        }

        private bool FilmeExists(int id)
        {
            return _context.Filmes.Any(e => e.Id == id);
        }
    }
}